# api/openapi.yml
openapi: 3.1.0
info:
  version: 0.0.1
  title: Brevity API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  x-logo:
    url: "https://redocly.github.io/openapi-template/logo.png"
    altText: OpenAPI example logo
...
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /v1/auth/me:
    get:
      security:
        - bearerAuth: []
      operationId: getCurrentUser
      description: Get the current user's information.
      tags:
        - auth
      responses:
        "200":
          description: "Successfully retrieved the current user."
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  username:
                    type: string
                  email:
                    type: string
                    format: email
                required:
                  - id
                  - username
                  - email
        "401":
          description: "Unauthorized."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...